Cardflow component cleanup
