+++ addon/styles/cardflow.css
@@ -603,7 +603,6 @@
 }
 
 .cardflow__sidebar-cards .boxel-card--collection {
-  min-width: 360px;
   min-height: 160px;
 }
